+++ src/accounting/Assets.h
@@ -5,20 +5,20 @@
 class Assets
 {
 public:
-    Assets(int equity, int debt=0); 
+    Assets(double equity, double debt=0); 
     Assets(Assets& source);
 
-    int get_equity() const;
-    int get_debt() const;
-    int total_assests() const;
-    int equity_to_debt_ratio() const;
+    double get_equity() const;
+    double get_debt() const;
+    double total_assests() const;
+    double equity_to_debt_ratio() const;
 
-    int add_equity(int add_to_equity);
-    int add_debt(int add_to_debt);
+    double add_equity(double add_to_equity);
+    double add_debt(double add_to_debt);
     
 private:
-    int m_equity;
-    int m_debt;
+    double m_equity;
+    double m_debt;
 };
 
 #endif // ASSESTS_H
